Separating GUIUtil Classes into their own separate files. This helps them stand out more to people looking through the code.
