+++ Assets/Packages/PrefsGUI/GUIUtil/Field.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: d115ffd4e8625074891cb8ee5dcbc8b3
+timeCreated: 1519369562
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
