+++ Assets/Packages/PrefsGUI/GUIUtil/Indent.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
+
+public static partial class GUIUtil
+{
+	public static void Indent(Action action) { Indent(1, action); }
+	public static void Indent(int level, Action action)
+	{
+		const int TAB = 20;
+		using (var h = new GUILayout.HorizontalScope())
+		{
+			GUILayout.Space(TAB * level);
+			using (var v = new GUILayout.VerticalScope())
+			{
+				action();
+			}
+		}
+	}
+}
