+++ Assets/Packages/PrefsGUI/GUIUtil/Slider.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 065bcde744068d647994901e0906c698
+timeCreated: 1519369684
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
